pyproject.toml
[tool.poetry]
name = "trips"
version = "0.1.0"
description = ""
authors = ["Mitchell Jindra <mitchjindra413@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
django = "^4.2"
psycopg2-binary = "^2.9.6"
python-dotenv = "^1.0.0"
djangorestframework = "^3.14.0"
django-cors-headers = "^3.14.0"
djangorestframework-simplejwt = "^5.2.2"
pylint = "^2.17.3"
serializers = "^0.2.4"


[tool.poetry.group.dev.dependencies]
autopep8 = "^2.0.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
